compilation fixes for non default wxUSE_XXX values (patch 662781)
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@18695 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/dc.h
@@ -602,7 +602,7 @@ public:
             float *descent = NULL, float *externalLeading = NULL,
             wxFont *theFont = NULL, bool use16bit = FALSE) const ;
     void GetSize(float* width, float* height) const { int w, h; GetSize(& w, & h); *width = w; *height = h; }
-    void GetSizeMM(float *width, float *height) const { long w, h; GetSizeMM(& w, & h); *width = (float) w; *height = (float) h; }
+    void GetSizeMM(float *width, float *height) const { int w, h; GetSizeMM(& w, & h); *width = (float) w; *height = (float) h; }
 
 #endif // WXWIN_COMPATIBILITY
